build: remove machine id
The file /etc/machine-id is supposed to be updated to a
unique value whenever the system boots for the first time.
However, when running within a container, this file always
contains the same id. This id can be picked up by telemetry,
so it is best to remove it unless we will do an operation to
replace the id when the container boots for the first time.

Signed-off-by: Elijah Zupancic <e.zupancic@f5.com>