build: define java logging in base image
The production log format for the Java apps is new line
delimited JSON. However, this format is difficult to parse
when developing or testing the application. Thus, in this
change we have moved the JSON log format definition into the
container build and away from the defaults that may be run
during development.

Signed-off-by: Elijah Zupancic <e.zupancic@f5.com>