fix: setup grpc to use host's DNS
This change adds the environment variable GRPC_DNS_RESOLVER=native
to the base Python Docker image. This is important because the
Python grpc library does not use the running host's DNS for
resolving DNS names. As such it causes problems when running on
localhost environments or in environments like kubernetes where
DNS is used for service location.

Signed-off-by: Elijah Zupancic <e.zupancic@f5.com>