build: fix java build failing in CI
This change upgrades Spring Sleuth to RC1 (a snapshot version)
and enables the Spring snapshots repository for the
telemetry-common module. Ideally, we do not want to be
depending on snapshots or milestone builds, but otel is so new
that we do not have much of a choice.

This change also makes running java-build a requirement for
java-test in the Makefile workflow.

Signed-off-by: Elijah Zupancic <e.zupancic@f5.com>