feat: add trace ids to Python logs
In order to add Open Telemetry trace ids to Python Flask
applications that are running in Gunicorn or in stand-alone
development mode, a few things needed to be in place.

1. Gunicorn must be configured dynamically.
2. Tracing must be configured dynamically.
3. Gunicorn and tracing configuration must be shared
   between applications.
4. The build and test process must be modified to support
   shared Python code between applications.

This change implements the above changes and also addresses
a potential problem with starting up Open Telemetry exporters
when running GUnicorn.

Signed-off-by: Elijah Zupancic <e.zupancic@f5.com>