refactor: gunicorn is now executed by bash script
Executing gunicorn by bash script gives us more
flexibility in how it is invoked. We can set default
values for log level, etc. Moreover, in this change
we are defining a default Docker CMD in the python
base image so that this does not need to be defined
in each inherited image.

Signed-off-by: Elijah Zupancic <e.zupancic@f5.com>